Fixed HttpContext extension class
